Rewritten KeyboardAPI using Django-Rest-Framework.
counter.js now expects to GET and POST/PUT to /api/keyboard/, and /api/keyboard/<id>/. These should be set in the project-level urls.py

#Features

##Clean URL structure

/keyboard/ - GET/POST to get or create new keyboards
/keyboard/<id> - GET/PUT/DELETE to get, update or delete existing keyboard

##Tidy dependencies

Javascript pertaining to the counter is now included within the cc_kapi app, rather than in main.

##Robust Test Suite

There is a Test Suite for the API. Reliant fairly heavily on integration tests, but functionality is covered.