Remove unnecessary license management
There is no need for complex license management. Users agree to a
forward-modifiable terms of service. This avoids having users on different
terms and forcing people to update their terms if changes are necessary.

Single URL for Terms which is served from a single URL for simplicity. At
signup users agree current terms, which may be modified in the future with
a given notice period.